Splitte ut deler av Familieralsjoner til ny SystemregistrertBarn komponent.
